[Add] #4 - 로그인 옵션 Fragment
- 로그인 옵션 Fragment UI 추가
- drawable이나 기타 이미지 등 이름 통일성 있게 수정해서 적용
