+++ app/src/main/res/drawable/login_logo.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="112dp"
+    android:height="130dp"
+    android:viewportWidth="112"
+    android:viewportHeight="130">
+  <path
+      android:pathData="M40.99,66.58C42.25,66.89 43.36,67.36 44.25,68.5C44.69,69.03 44.88,69.77 44.75,70.56C44.62,71.36 44.14,72.46 43.6,72.9C42.24,73.91 39.37,73.35 39.25,73.32L7.48,68.33C7.13,68.27 6.08,68.17 5.73,68.11C4.05,67.91 2,67.71 0.85,66.41C0.21,65.61 -0.11,64.68 0.04,63.77C0.22,62.62 1.21,61.32 2.44,61.05C4.08,60.73 5.74,61.1 7.44,61.31L7.73,61.35C8.89,61.53 10.69,61.76 11.91,61.95C11.93,61.48 11.98,61.14 11.98,61.14L12.84,55.83C12.84,55.83 13.12,54.06 14.7,53.13C11.97,52 9.59,50.22 8.16,47.3C6.81,44.69 6.79,41.88 7.25,39.02C8.21,33.14 12.34,29.58 18.13,28.67C21.97,28.05 25.95,28.38 29.72,29.27C32.81,29.99 36.08,30.68 38.76,32.45C39.98,33.28 41.07,34.28 41.99,35.59C44.25,38.88 44.56,42.43 43.66,46.16C42.24,52.02 38.27,54.91 33.44,55.79C34.18,56.37 34.69,57.21 34.58,58.3C34.44,59.51 34.05,62.31 33.55,65.39L35.86,65.76C37.24,66.04 39.11,66.15 40.32,66.46C40.54,66.56 40.78,66.54 41.01,66.57L40.99,66.58ZM36.27,44.42C36.79,40.46 34.43,38.23 30.86,37.19C26.6,35.99 17.28,33.41 15.37,39.37C14.26,42.95 15.25,46.37 20.66,47.75C23.32,48.47 26.1,49.31 28.9,49.23C31.65,49.13 34.77,48.51 35.92,45.82C36.12,45.39 36.25,44.94 36.27,44.41V44.42ZM19.84,54.65C20,55.09 20.11,55.52 20.09,56.03C19.96,57.24 19.56,60.04 19.06,63.13L25.19,64.09L26.35,64.27C26.37,63.75 26.43,63.4 26.43,63.4L27.29,58.1C27.29,58.1 27.48,56.9 28.4,55.99C27.7,55.94 27,55.83 26.31,55.72C24.28,55.4 22.08,55.11 19.84,54.64V54.65Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M44.04,114.51C45.33,114.37 46.54,114.43 47.75,115.12C48.37,115.53 48.8,116.15 48.97,116.94C49.13,117.73 49.08,118.92 48.73,119.53C47.76,120.97 44.93,121.42 44.81,121.44L12.66,127.89C12.31,127.95 11.92,128.09 11.52,128.18C9.92,128.55 7.88,129.09 6.4,128.26C5.51,127.73 4.88,126.98 4.69,126.07C4.45,124.93 4.91,123.37 5.97,122.69C7.41,121.81 9.1,121.59 10.75,121.2L11.03,121.15C14.7,120.36 18.48,119.6 22.28,118.89L21,112.95C21,112.95 20.6,111.03 21.94,109.64C23.6,107.84 27.7,108.02 28.27,110.79L29.69,117.4C33.3,116.62 35.88,116.1 39.46,115.44C40.85,115.22 42.14,114.79 43.37,114.66C43.61,114.66 43.82,114.57 44.05,114.52L44.04,114.51ZM31.06,82.03C33.82,81.47 37.39,81.41 37.92,83.95C38.69,87.69 35.48,88.33 32.85,88.68C30.3,89.08 28.36,89.53 25.4,90.23L25.77,91.75C25.9,92.37 26.03,93.29 26.27,93.89C26.67,94.7 27.53,95.11 28.24,95.62C30.98,97.54 34.83,99.43 37.72,101.32C39.46,102.51 42.05,104.05 41.69,106.31C41.16,109.3 37.29,109.08 35.2,108.2C30.07,105.99 27.38,104.22 23.68,101.96C23.49,101.89 20.71,105.21 19.55,106.56C18.53,107.71 16.1,110.44 14.04,112.33C11.98,114.28 9.63,115.64 7.57,113.74C5,111.13 7.54,108.68 9.27,106.62C11.87,103.56 14.68,100.35 16.75,97.34C17.57,96.11 17.52,95.59 17.27,94.34L16.76,91.85L8.14,93.57C8.14,93.57 3.22,95.21 2.22,91.81C1.7,90.14 3.15,87.9 5.64,87.23C6.15,87.07 5.1,87.27 10.74,86.15L31.05,82.02L31.06,82.03Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M60.34,95L62.97,75.14C63.14,73.83 63.32,72.46 63.37,71.07C63.36,70.62 63.41,69.68 62.93,69.42C62.47,69.11 61.82,69.15 61.26,69.01C60.63,68.87 59.98,68.92 59.35,68.77L56.03,68.34C54.47,68.2 52.79,67.86 51.22,67.72C50.51,67.68 49.88,67.55 49.17,67.45C48.11,67.31 46.92,67.16 46.07,66.48C44.83,65.49 44.66,64.07 44.83,62.76C44.87,61.94 45.19,61.16 45.71,60.46C45.98,60.06 46.31,59.71 46.78,59.52C47.33,59.15 48.42,58.97 49.09,58.93C51.13,58.75 53.37,59.1 55.42,59.37L67.72,60.95C68.9,61.23 70.37,61.55 71.45,62.07C72.12,62.41 72.92,62.96 73.26,63.58C73.96,64.87 73.81,65.99 73.76,67.39C73.67,69.73 73.3,72.03 72.98,74.34L70.27,94.83C70.25,95.01 69.69,100.22 69.21,102.82C68.84,104.49 68.14,106.5 65.56,107C64.67,107.2 63.8,107.21 63.1,107.13C61.89,106.98 60.53,106.36 60.17,105.35C59.62,104.01 59.41,102.4 59.62,100.77L59.99,98.02C60.06,97.01 60.19,96.01 60.32,95.01L60.34,95Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M93.35,44.69C92.6,44.69 91.82,44.66 91.03,44.59C84.51,44.29 79.74,42.13 76.89,38.16C70.61,29.41 75.22,14.36 77.69,6.27C77.91,5.58 78.1,4.95 78.27,4.38C78.89,2.28 80.39,0.67 82.22,0.16C83.66,-0.24 85.09,0.11 86.24,1.14L86.36,1.25C87.15,1.98 87.56,3.18 87.46,4.42L85.86,23.53L88.45,23.75L90.09,2.86C90.09,2.78 90.18,1.74 91.13,0.92C91.94,0.22 93.11,-0.08 94.6,0.05C97.91,0.31 98.86,2.39 98.9,3.45V3.51L97.05,24.44L99.63,24.66L101.2,5.95C101.34,4.26 102.23,2.83 103.52,2.19C105.29,1.31 107.15,1.61 108.78,3.01C109.71,3.82 110.04,4.97 110.22,5.88L110.31,6.33C114.29,25.89 110.52,35.13 106.64,39.44C103.51,42.92 99.04,44.69 93.36,44.69H93.35ZM83.36,1.97C83.16,1.97 82.95,2.01 82.75,2.06C81.61,2.38 80.6,3.5 80.18,4.94C80,5.5 79.81,6.15 79.6,6.85C77.22,14.61 72.8,29.07 78.52,37.04C80.99,40.49 85.24,42.38 91.15,42.65C97.33,43.13 102.02,41.61 105.15,38.14C108.72,34.17 112.18,25.49 108.36,6.72L108.26,6.26C108.09,5.37 107.84,4.83 107.46,4.5C105.95,3.19 104.79,3.77 104.4,3.95C103.75,4.27 103.25,5.15 103.17,6.13L101.45,26.79L94.88,26.25L96.89,3.5C96.86,3.28 96.56,2.2 94.42,2.02C92.3,1.84 92.09,2.83 92.07,3.02L90.26,25.88L83.69,25.34L85.45,4.27C85.5,3.63 85.31,3 84.96,2.67C84.44,2.2 83.9,1.97 83.35,1.97H83.36Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M94.77,63.35L84.47,62.52L86.17,42.27L96.46,43.11L94.77,63.35ZM86.55,60.78L93.01,61.31L94.39,44.84L87.92,44.31L86.55,60.78Z"
+      android:fillColor="#E9E9E9"/>
+  <path
+      android:pathData="M89.25,129.75C89.14,129.75 89.03,129.75 88.92,129.74L79.23,128.95C78.21,128.86 77.28,128.39 76.62,127.62C75.96,126.85 75.63,125.86 75.72,124.86L80.79,64.07C80.88,63.06 81.36,62.14 82.14,61.49C82.92,60.84 83.92,60.53 84.94,60.6L94.63,61.39C96.75,61.56 98.32,63.4 98.15,65.48L93.07,126.27C92.98,127.28 92.5,128.2 91.72,128.85C91.02,129.44 90.15,129.75 89.24,129.75H89.25ZM84.63,62.58C84.19,62.58 83.78,62.73 83.45,63.01C83.08,63.32 82.86,63.76 82.81,64.23L77.73,125.02C77.69,125.5 77.84,125.96 78.16,126.32C78.47,126.69 78.92,126.91 79.4,126.96L89.09,127.75C89.58,127.79 90.04,127.64 90.41,127.33C90.78,127.02 91.01,126.58 91.05,126.11L96.13,65.32C96.22,64.34 95.47,63.46 94.46,63.38L84.77,62.59C84.77,62.59 84.67,62.59 84.61,62.59L84.63,62.58Z"
+      android:fillColor="#E9E9E9"/>
+</vector>
