+++ app/src/main/res/drawable/login_opt_kakao_back.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="313dp"
+    android:height="46dp"
+    android:viewportWidth="313"
+    android:viewportHeight="46">
+  <path
+      android:pathData="M20,0L293,0A20,20 0,0 1,313 20L313,26A20,20 0,0 1,293 46L20,46A20,20 0,0 1,0 26L0,20A20,20 0,0 1,20 0z"
+      android:fillColor="#FEE500"/>
+</vector>
