+++ app/src/main/res/drawable/login_opt_naver_back.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="313dp"
+    android:height="47dp"
+    android:viewportWidth="313"
+    android:viewportHeight="47">
+  <path
+      android:pathData="M20,-0L293,-0A20,20 0,0 1,313 20L313,27A20,20 0,0 1,293 47L20,47A20,20 0,0 1,0 27L0,20A20,20 0,0 1,20 -0z"
+      android:fillColor="#03C75A"/>
+</vector>
